Split scar and weakness sections
